revert: prior code changes to isolate bug
This commit reverts the following commit.

* d91d943c1b6fae15274e1a8fb24252c0325068a8

Starting 04/05/2024, deployments of this service started to fail due to failing migrations on the wagtail core model in the wagtail application. We are reverting the code changes up until the issue began to rule out an issue in any of our Python requirements upgrade, namely the upgrade of wagtail from 6.0.1 to 6.0.2.

Revert "chore: Updating Python Requirements (#312)"

This reverts commit d91d943c1b6fae15274e1a8fb24252c0325068a8.