revert: prior code changes to isolate bug
This commit reverts the following commit.

* e8883bdbfb271cbf1810b8b5eced7723233be16d

Starting 04/05/2024, deployments of this service started to fail due to failing migrations on the wagtail core model in the wagtail application. We are reverting the code changes up until the issue began to rule out an issue in any of our Python requirements upgrade, namely the upgrade of wagtail from 6.0.1 to 6.0.2.

Revert "chore: Updating Python Requirements"

This reverts commit e8883bdbfb271cbf1810b8b5eced7723233be16d.